Boardクラスにロジックを移植したためcore.tsを削除
